+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removed cannon-es library support completely from editor
 - Removed Gamepad and TizenKeyboard functionality from editor
 - Removed cordova related code from editor
+- Removed jszip related code from editor
 
 ### V0.108